app/requirements.txt
gym[classic_control]
pyglet
numpy
matplotlib
tensorflow
tqdm
keras
pyyaml